Add Preprocessing tests

+++ src/Input.jl
@@ -21,7 +21,7 @@ function read_xdf(infile::String)
     # Find the stream with "RR" in the name
     rr_idx = findfirst(x -> occursin("RR", x["name"]), data)
     #TODO: time management with the XDF timestamps
-    return Float32.(data[rr_idx]["data"])
+    return Float64.(data[rr_idx]["data"])
 end
 
 """
@@ -38,7 +38,7 @@ Returns:
 """
 function read_txt(infile::String)
     a=read(open(infile, "r"), String)
-    return parse.(Float32, filter!(e->e!="", split(a, r"[^\d.]")))
+    return parse.(Float64, filter!(e->e!="", split(a, r"[^\d.]")))
 end
 
 """
